chore: initDatabase() UserService로 위치 변경
