+++ src/main/java/makar/dev/controller/StationController.java
@@ -1,8 +1,6 @@
 package makar.dev.controller;
 
 import lombok.RequiredArgsConstructor;
-import makar.dev.common.response.ApiResponse;
-import makar.dev.common.status.SuccessStatus;
 import makar.dev.service.StationService;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -14,9 +12,5 @@ import org.springframework.web.bind.annotation.RestController;
 public class StationController {
     private final StationService stationService;
 
-    @GetMapping("/init")
-    ApiResponse initStation(){
-        stationService.initDatabase();
-        return ApiResponse.SuccessResponse(SuccessStatus._OK);
-    }
+
 }
