+++ src/main/java/makar/dev/controller/UserController.java
@@ -1,7 +1,10 @@
 package makar.dev.controller;
 
 import lombok.RequiredArgsConstructor;
+import makar.dev.common.response.ApiResponse;
+import makar.dev.common.status.SuccessStatus;
 import makar.dev.service.UserService;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -10,5 +13,11 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/v1")
 public class UserController {
     private final UserService userService;
+
+    @GetMapping("/init")
+    ApiResponse initStation(){
+        userService.initDatabase();
+        return ApiResponse.SuccessResponse(SuccessStatus._OK);
+    }
 
 }
